+++ Billing/app/Dockerfile
@@ -5,6 +5,6 @@ WORKDIR /app
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY app.py .
+COPY app.py routes.py db.py .
 
 CMD ["python", "app.py"]
